TODO-RELEASE: push the remaining ones to next release
As we're closing in on the release, I give up on the remaining ones but
I leave them in here for now to try to fix for next release.

I removed the 281 issue about warnings from the statical analyzer scans,
as they seem to be mostly false positives at this point.